requirements.txt
Django>=4.2
djangorestframework
mongoengine
python-decouple